bootstrap_monitor_cluster: simplify (and make work) mon_relation: cast int to int so that i can int while i int

+++ hooks/hooks.py
@@ -81,30 +81,33 @@ def bootstrap_monitor_cluster():
     secret = utils.config_get('monitor-secret')
     keyring = "/var/lib/ceph/tmp/%s.mon.keyring" % hostname
 
-    try:
-        with os.fdopen(os.open(done, os.O_WRONLY | os.O_CREAT | os.O_EXCL)) as d:
-            try:
-                subprocess.check_call(['ceph-authtool', keyring,
-                                       '--create-keyring' '--name=mon.'
-                                       "--add-key=%s" % secret,
-                                       '--cap', 'mon', 'allow *'])
-
-                subprocess.check_call(['ceph-mon', '--mkfs',
-                                       '-i', hostname,
-                                       '--keyring', keyring])
-            except:
-                os.unlink(done)
-                os.unlink(keyring)
-                raise
-        subprocess.check_call(['start', 'ceph-mon-all-starter'])
-    except OSError:
+    if os.path.exists(done):
         utils.juju_log('INFO', 'bootstrap_monitor_cluster: mon already initialized, getting on with life.')
+    else:
+        try:
+            subprocess.check_call(['ceph-authtool', keyring,
+                                   '--create-keyring', '--name=mon.',
+                                   "--add-key=%s" % secret,
+                                   '--cap', 'mon', 'allow *'])
+
+            subprocess.check_call(['ceph-mon', '--mkfs',
+                                   '-i', hostname,
+                                   '--keyring', keyring])
+
+            with open(done, 'w'):
+                pass
+
+            subprocess.check_call(['start', 'ceph-mon-all-starter'])
+        except:
+            raise
+        finally:
+            os.unlink(keyring)
 
 def mon_relation():
     utils.juju_log('INFO', 'Begin mon-relation hook.')
     emit_cephconf()
 
-    moncount = utils.config_get('monitor-count')
+    moncount = int(utils.config_get('monitor-count'))
     if len(get_mon_hosts()) == moncount:
         bootstrap_monitor_cluster()
     else:
